+++ market-service/src/main/java/rs/edu/raf/banka1/model/dtos/ListingHistoryDto.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class ListingHistoryDto {
     private Long listingHistoryId;
     private String ticker;
-    private Integer date;
+    private long date;
     private Double price;
     private Double high;
     private Double low;
